WIP: Add cloudwatch agent
Add Cloudwatch Metrics agent, written in Go.

This is a MVP -- the code is not very idiomatic in style, not tested
in certain areas (particularly around Task ID detection), and is
generally ugly and brittle.  But it collects metrics from Envoy and
posts them to CloudWatch, such as it is.

I'll be refactoring it shortly to be prettier and more functional.