CloudWatch agent: eliminate ECS task ID dimension
CloudWatch metrics cannot perform metric aggregation across multiple
dimensions, so supplying the task ID as a dimension makes things a mess
and doesn't provide the sort of value we want it to.  There's more
granularity, but most of the value is in the aggregated metrics.

Eliminate the dimension and pray that CloudWatch doesn't replace values
from two different senders that have the same dimensions and timestamps.